# API Layer/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Smart App Group C - API Layer
  description: |-
    This is the API used for internal communication between services and by external applications and tools to allow data access.
    Made by the integration team: Andrea Alberti, Anna Francesca Montagnoli, Matteo Simone
  version: 1.1.1
servers:
- url: /
...
paths:
...
  /machine:
    get:
      tags:
      - machine data
      summary: Gets machine related informations
      description: "Returns machine informations including status according to optional\
        \ filters. The query can be restricted over machine ids, machine types, production\
        \ lines, factories and machine status. A non specified parameter implies no\
        \ selection will be performed over it. For efficiency purposes, if status\
        \ data is not required, refer to /machine/summary instead."
      operationId: machine_get
      parameters:
      - name: machines
        in: header
        description: Specific machines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineIDFilter'
      - name: machineTypes
        in: header
        description: Specific machine types to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: string
            example: Laser Welding
        example:
        - Laser Welding
      - name: productionLines
        in: header
        description: Specific production lines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: integer
            example: 2
        example:
        - 1
      - name: factories
        in: header
        description: Specific factories to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: string
            example: Factory X
        example:
        - Factory X
      - name: machineStatus
        in: header
        description: Specific machine status to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineStatus'
      responses:
        "200":
          description: Succesful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MachineCompleteInfo'
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
    post:
      tags:
      - machine data
      summary: Adds a new machine to the system
      description: Adds a new machine to the system by providing its informations.
        The created machine id will be returned in case of a successful request.
      operationId: machine_post
      parameters:
      - name: name
        in: header
        description: The name of the machine
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Laser Welding Machine 10
      - name: productionLine
        in: header
        description: The production line of the machine
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          example: 2
      - name: factory
        in: header
        description: The factory of the machine
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Factory X
      - name: machineType
        in: header
        description: The machine type of the machine
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Laser Welding
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineNotFoundError'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
  /machine/summary:
    get:
      tags:
      - machine data
      summary: Gets basic machine informations
      description: "Returns machine informations according to optional filters. The\
        \ query can be restricted over machine ids, machine types, production lines\
        \ and factories. A non specified parameter implies no selection will be performed\
        \ over it."
      operationId: machine_summary_get
      parameters:
      - name: machines
        in: header
        description: Specific machines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineIDFilter'
      - name: machineTypes
        in: header
        description: Specific machine types to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: string
            example: Laser Welding
        example:
        - Laser Welding
      - name: productionLines
        in: header
        description: Specific production lines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: integer
            example: 2
        example:
        - 1
      - name: factories
        in: header
        description: Specific factories to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          type: array
          items:
            type: string
            example: Factory X
        example:
        - Factory X
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MachineBasicInfo'
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
  /machine/status:
    get:
      tags:
      - machine data
      summary: Gets status informations of machines
      description: "Returns the status of all machines unless a list of machine ids\
        \ is provided as filter, in which case only the requested machines status\
        \ will be returned. The status of a machine indicates whether sensor data\
        \ for it has not been collected for 30 hours."
      operationId: machine_status_get
      parameters:
      - name: machines
        in: header
        description: Specific machines to restrict the query to
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/MachineIDFilter'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200'
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
  /machine/{id}:
    get:
      tags:
      - machine data
      summary: Returns informations of a given machine
      description: "Returns all informations of a machine given its id(including machine\
        \ status, see /machine/status for details)."
      operationId: machine_id_get
      parameters:
      - name: id
        in: path
        description: ID of the resource
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/ID'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineCompleteInfo'
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineNotFoundError'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
    put:
      tags:
      - machine data
      summary: Updates machine data
      description: "Modifies the name, production line, factory and/or machine type\
        \ of a given machine. If a parameter is not provided, the original machine\
        \ data for it will be kept"
      operationId: machine_id_put
      parameters:
      - name: id
        in: path
        description: ID of the resource
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/ID'
      - name: name
        in: header
        description: The updated name of the machine
        required: false
        style: simple
        explode: false
        schema:
          type: string
          example: Laser Welding Machine 10
      - name: productionLine
        in: header
        description: The updated production line of the machine
        required: false
        style: simple
        explode: false
        schema:
          type: integer
          example: 2
      - name: factory
        in: header
        description: The updated factory of the machine
        required: false
        style: simple
        explode: false
        schema:
          type: string
          example: Factory X
      - name: machineType
        in: header
        description: The updated machine type of the machine
        required: false
        style: simple
        explode: false
        schema:
          type: string
          example: Laser welding
      responses:
        "200":
          description: Success Message
          content:
            application/json:
              schema:
                type: string
                example: Success
                x-content-type: application/json
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid parameters
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineNotFoundError'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
    delete:
      tags:
      - machine data
      summary: Deletes machine informations
      description: Deletes informations of a given machine
      operationId: machine_delete
      parameters:
      - name: id
        in: path
        description: ID of the resource
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/ID'
      responses:
        "200":
          description: Success Message
          content:
            application/json:
              schema:
                type: string
                example: Success
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineNotFoundError'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
  /machine/{id}/KPIList:
    get:
      tags:
      - machine data
      summary: Returns machine KPI list
      description: Returns the list of KPI supported by the machine according to its
        type. A KPI is considered as supported if the machine directly produces it
        or every KPI needed for its computation are supported by the machine.
      operationId: machine_id_kpi_list_get
      parameters:
      - name: id
        in: path
        description: ID of the resource
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/ID'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KPIID'
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineNotFoundError'
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.machine_data_controller
...
components:
  schemas:
...
    MachineNotFoundError:
      type: string
      description: Machine not found error message
      example: Machine not found
...
    ID:
      type: integer
      description: Numerical unique identifier of the resource
      format: int64
      example: 1
    KPIID:
      type: string
      description: A unique identifier of a KPI
      example: good_cycles
...
    MachineIDFilter:
      type: array
      description: List of machine identifiers
      example:
      - 1
      - 2
      items:
        $ref: '#/components/schemas/ID'
    MachineCompleteInfo:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/MachineStatus'
      description: Holds informations of a machine
      example:
        status: idle
      allOf:
      - $ref: '#/components/schemas/MachineBasicInfo'
    MachineBasicInfo:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        machineType:
          type: string
          example: Laser cutter
        productionLine:
          type: integer
          example: 2
        factory:
          type: string
          example: Ghezzano Plant 1
        name:
          type: string
          example: Laser Welding Machine 2
      description: Holds basic informations of a machine
      example:
        factory: Ghezzano Plant 1
        productionLine: 2
        name: Laser Welding Machine 2
        id: 1
        machineType: Laser cutter
    MachineStatus:
      type: string
      description: "The status of a machine. A machine is considered operational if\
        \ and only if sensor data for it has been received in the past 30 hours, otherwise\
        \ it is conidered idle."
      example: idle
      enum:
      - operational
      - idle
...
    inline_response_200:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        status:
          $ref: '#/components/schemas/MachineStatus'
      example:
        id: 1
        status: idle
...
  securitySchemes:
    api_token_auth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_api_token_auth
    machine_auth:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:aiquery: Get ai query responses
            write:aiquery: Request new ai queries
            read:data: Access raw data
        authorizationCode:
          authorizationUrl: https://SA2024C.com/oauth/authorize
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:data: Access raw data
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_machine_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_machine_auth